// src/api/model/registerProductPhotoRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Forgather API
 * 당신을 위한 순간, 흩어지지 않게. Forgather
 * OpenAPI spec version: 2.0.0
 */

/**
 * 작품 사진 목록
 */
export interface RegisterProductPhotoRequest {
  originalName?: string;
  uploadFileName?: string;
  capacity?: number;
}